Split the DTOs and the address client
